backend/requirements.txt
# --- Основные зависимости ---
fastapi==0.110.0
uvicorn[standard]==0.29.0

# --- Работа с БД ---
SQLAlchemy==2.0.30
asyncpg==0.29.0  # Асинхронный драйвер для PostgreSQL
psycopg2-binary
alembic==1.13.1
python-multipart>=0.0.5

# --- Валидация и сериализация ---
pydantic==2.7.1
pydantic-settings==2.2.1

# --- Асинхронные задачи ---
celery==5.3.6
redis==5.0.3
pandas
fastapi_mail

# --- Обработка CORS (если API вызывается с фронтенда) ---
fastapi[all]==0.110.0  # Можно убрать, если хотите минимальный набор

# --- Тестирование ---
pytest==8.2.1
httpx==0.27.0
pytest-asyncio==0.23.6
coverage==7.5.1

# --- Утилиты разработки ---
python-dotenv==1.0.1
bcrypt
python-jose

openpyxl==3.1.2        # генерация Excel
matplotlib==3.9.0      # графики
reportlab==4.1.0       # PDF-отчёты
